Refactor #17 - Move di module to internal pacakage
