+++ app/src/main/java/com/nlab/reminder/domain/internal/di/feature/home/HomeEntryPointInitModule.kt
@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.common.effect.message.navigation.di
+package com.nlab.reminder.domain.internal.di.feature.home
 
 import androidx.fragment.app.Fragment
 import com.nlab.reminder.core.effect.message.navigation.android.NavigationMediator
 import com.nlab.reminder.core.effect.message.navigation.android.fragment.util.toReceiver
+import com.nlab.reminder.core.entrypoint.EntryPointInit
 import com.nlab.reminder.core.entrypoint.util.DefaultEntryPointInit
 import com.nlab.reminder.core.entrypoint.util.EntryBlock
+import com.nlab.reminder.core.entrypoint.util.concat
+import com.nlab.reminder.core.entrypoint.util.plus
+import com.nlab.reminder.domain.feature.home.*
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -32,14 +36,17 @@ import dagger.hilt.android.scopes.FragmentScoped
  */
 @Module
 @InstallIn(FragmentComponent::class)
-class EntryPointInitFragmentModule {
+class HomeEntryPointInitModule {
+    @HomeScope
     @FragmentScoped
     @Provides
     fun provideEntryPointInit(
         fragment: Fragment,
-        @FragmentScoped navigationMediator: NavigationMediator
-    ): DefaultEntryPointInit = DefaultEntryPointInit(
+        defaultEntryPointInit: DefaultEntryPointInit,
+        @HomeScope navigationMediator: NavigationMediator,
+        @HomeScope blocks: Set<@JvmSuppressWildcards EntryBlock>
+    ): EntryPointInit = defaultEntryPointInit.copy(
         navigationEffectReceiver = navigationMediator.toReceiver(fragment),
-        block = EntryBlock {}
+        block = defaultEntryPointInit.block + blocks.concat()
     )
 }
